+++ spec/api_spec.rb
@@ -12,7 +12,7 @@ RSpec.describe All3DP::API do
       )
     }
     it "returns a response" do
-      stub_request(:post, "https://printing-engine.all3dp.com/configuration")
+      stub_request(:post, "https://api.craftcloud3d.com/configuration")
         .with(
           body: '{"items":[{"modelUrl":"http://example.org/file.stl"}]}',
           headers: {
@@ -39,7 +39,7 @@ RSpec.describe All3DP::API do
       }
 
       it "raises an exception" do
-        stub_request(:post, "https://printing-engine.all3dp.com/configuration")
+        stub_request(:post, "https://api.craftcloud3d.com/configuration")
           .to_return(status: 502, body: body)
 
         expect { result }
@@ -49,7 +49,7 @@ RSpec.describe All3DP::API do
 
     context "when 503" do
       it "raises an exception" do
-        stub_request(:post, "https://printing-engine.all3dp.com/configuration")
+        stub_request(:post, "https://api.craftcloud3d.com/configuration")
           .to_return(status: 503, body: "")
 
         expect { result }
@@ -59,7 +59,7 @@ RSpec.describe All3DP::API do
 
     context "when 504" do
       it "raises an exception" do
-        stub_request(:post, "https://printing-engine.all3dp.com/configuration")
+        stub_request(:post, "https://api.craftcloud3d.com/configuration")
           .to_return(status: 504, body: "")
 
         expect { result }.to raise_error(All3DP::API::GatewayTimeoutError, "")
@@ -71,7 +71,7 @@ RSpec.describe All3DP::API do
       let(:body) { "meh" }
 
       it "raises an exception" do
-        stub_request(:post, "https://printing-engine.all3dp.com/configuration")
+        stub_request(:post, "https://api.craftcloud3d.com/configuration")
           .to_return(status: status_code, body: body)
 
         expect { result }